+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Harvest control logic wasn't working properly for growing zones since I had misunderstood the logic of "allow cut" on growing zones.
 
 ## [0.6.0] - 2023-01-04
 ### Added